cleanup main; consolidate code, set server addrs w/ env
