+++ crates/server/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod config;
 pub mod error;
 pub mod log_layer;
 pub mod rpc;
