+++ crates/server/src/main.rs
@@ -1,41 +1,30 @@
 use std::sync::Arc;
 
-use contender_server::log_layer::{new_log_sinks, SessionLogRouter};
+use contender_server::config::{init_tracing, load_server_config};
 use contender_server::rpc::{ContenderRpcServer as _, ContenderServer};
 use contender_server::sessions::ContenderSessionCache;
 use contender_server::sse::sse_router;
 use jsonrpsee::server::{Server, ServerHandle};
 use tokio::sync::RwLock;
+use tokio::task::JoinHandle;
 use tracing::info;
-use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, EnvFilter};
-
-async fn start_rpc_server(
-    sessions: Arc<RwLock<ContenderSessionCache>>,
-) -> std::io::Result<ServerHandle> {
-    let addr = "127.0.0.1:3000";
-    let server = Server::builder().build(addr).await?;
-    let module = ContenderServer::new(sessions).into_rpc();
-    let handle = server.start(module);
-
-    info!("JSON-RPC server listening on {addr}");
-    Ok(handle)
-}
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let log_sinks = new_log_sinks();
+    // initialize logging w/ a custom layer that pipes logs to session-specific broadcast channels
+    let log_sinks = init_tracing();
 
-    init_tracing(log_sinks.clone());
+    // load server config
+    let config = load_server_config();
 
+    // shared session cache
     let sessions = Arc::new(RwLock::new(ContenderSessionCache::new(log_sinks)));
 
-    let handle = start_rpc_server(sessions.clone()).await?;
+    // RPC server for session management and log subscription
+    let handle = start_rpc_server(sessions.clone(), &config.rpc_addr).await?;
 
-    // SSE endpoint for log streaming (port 3001)
-    let sse_app = sse_router(sessions);
-    let sse_listener = tokio::net::TcpListener::bind("127.0.0.1:3001").await?;
-    info!("SSE server listening on 127.0.0.1:3001");
-    let sse_handle = tokio::spawn(async move { axum::serve(sse_listener, sse_app).await });
+    // SSE endpoint for log streaming
+    let sse_handle = start_sse_server(sessions, &config.sse_addr).await?;
 
     tokio::select! {
         _ = tokio::signal::ctrl_c() => {
@@ -52,19 +41,32 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
 
-fn init_tracing(log_sinks: contender_server::log_layer::SessionLogSinks) {
-    let filter = EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new("info"));
+/// Starts a JSON-RPC HTTP server for managing contender sessions,
+/// which includes a websocket server for subscribing to session logs.
+///
+/// Returns a handle to the RPC server; awaiting `.stopped()` on this handle will wait until the server shuts down.
+async fn start_rpc_server(
+    sessions: Arc<RwLock<ContenderSessionCache>>,
+    addr: &str,
+) -> std::io::Result<ServerHandle> {
+    let server = Server::builder().build(addr).await?;
+    let module = ContenderServer::new(sessions).into_rpc();
+    let handle = server.start(module);
 
-    let fmt_layer = tracing_subscriber::fmt::layer()
-        .with_ansi(true)
-        .with_target(true)
-        .with_line_number(true);
+    info!("JSON-RPC server listening on {addr}");
+    Ok(handle)
+}
 
-    let session_layer = SessionLogRouter::new(log_sinks);
-
-    tracing_subscriber::registry()
-        .with(filter)
-        .with(fmt_layer)
-        .with(session_layer)
-        .init();
+/// Starts a simple SSE server that serves session logs at `/logs/:session_id`.
+///
+/// Returns a handle to the server task; awaiting this handle will wait until the server shuts down.
+async fn start_sse_server(
+    sessions: Arc<RwLock<ContenderSessionCache>>,
+    addr: &str,
+) -> std::io::Result<JoinHandle<std::io::Result<()>>> {
+    let sse_app = sse_router(sessions);
+    let sse_listener = tokio::net::TcpListener::bind(addr).await?;
+    info!("SSE server listening on {addr}");
+    let sse_handle = tokio::spawn(async move { axum::serve(sse_listener, sse_app).await });
+    Ok(sse_handle)
 }
